requirements/base.txt
asgiref==3.6.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==3.2.16
django-addanother==2.2.2
django-environ==0.9.0
django-filter==21.1
django-model-utils==4.3.1
django-modeltranslation==0.18.7
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-gis==1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.9.0
flake8==6.0.0
identify==2.5.18
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
nodeenv==1.7.0
packaging==23.0
Pillow==9.4.0
pip==22.3.1
platformdirs==3.0.0
pre-commit==3.0.4
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==1.7.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
setuptools==65.5.1
sqlparse==0.4.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
velait-api==0.4.2
virtualenv==20.19.0
wheel==0.38.4
django-ckeditor==6.7.0
